src/third_party/blink/web_tests/http/tests/origin_trials/webexposed/legacy-performance-memory-counters-disabled.html

<!DOCTYPE html>
<meta charset="utf-8">

<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script>

test(t => {
  assert_true(
      performance.memory.usedJSHeapSizeWithoutExternalMemory === undefined);
  assert_true(
      performance.memory.totalJSHeapSizeWithoutExternalMemory === undefined);
}, "Legacy performance memory counters are disabled without Origin Trial.");

</script>
